refactor(shared): move repository interfaces from Application to Domain layer (US-011)
Repository interfaces define the persistence contract the domain needs,
so they belong next to their aggregate in the Domain layer per DDD
port/adapter conventions. Service interfaces stay in Application.
