--- src/Yumney.Recipes.Application/Interfaces/IRecipeRepository.cs
@@ -1,5 +1,0 @@
-namespace Yumney.Recipes.Application.Interfaces;
-
-public interface IRecipeRepository
-{
-}
