+++ src/Yumney.Recipes.Domain/Recipe/IRecipeRepository.cs
@@ -1,0 +1,5 @@
+namespace Yumney.Recipes.Domain.Recipe;
+
+public interface IRecipeRepository
+{
+}
