--- src/Yumney.Shopping.Application/Interfaces/IShoppingListRepository.cs
@@ -1,5 +1,0 @@
-namespace Yumney.Shopping.Application.Interfaces;
-
-public interface IShoppingListRepository
-{
-}
