+++ src/Yumney.Shopping.Domain/IShoppingListRepository.cs
@@ -1,0 +1,5 @@
+namespace Yumney.Shopping.Domain;
+
+public interface IShoppingListRepository
+{
+}
